no real changes: adjust the header double-inclusion guard symbols
git-svn-id: https://usbpicprog.svn.sourceforge.net/svnroot/usbpicprog@556 29735491-9d32-0410-b17f-ed4c7e910dea

+++ trunk/upp_wx/src/uppmainwindow.cpp
@@ -132,7 +132,7 @@ UppMainWindow::UppMainWindow(wxWindow* parent, wxWindowID id)
     int lastPic;
     if (!m_hardware->connected() &&
         pCfg->Read(wxT("SelectedPIC"), &lastPic) &&
-        lastPic >= 0 && lastPic < m_arrPICName.size())
+        lastPic >= 0 && lastPic < (int)m_arrPICName.size())
     {
         m_picType=PicType(m_arrPICName[lastPic]);
